requirements.txt
numpy
pandas
joblib

scikit-learn

kafka-python

psycopg2-binary

streamlit
plotly

